chore: Add typescript
