--- cypress/support/index.js
@@ -1,2 +1,0 @@
-// adds cy.api command
-import '../../support'
