+++ cypress/support/index.ts
@@ -1,0 +1,2 @@
+// adds cy.api command
+import '../../dist/support'
