+++ package.json
@@ -2,11 +2,10 @@
   "name": "@bahmutov/cy-api",
   "version": "0.0.0-development",
   "description": "Custom command cy.api",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
-    "support.js",
-    "index.d.ts"
+    "/dist/*"
   ],
   "scripts": {
     "test": "cypress run",
@@ -16,7 +15,9 @@
     "dev": "start-test 3003 cy:open",
     "semantic-release": "semantic-release",
     "stop-only": "stop-only --folder cypress",
-    "warn-only": "stop-only --warn --folder cypress"
+    "warn-only": "stop-only --warn --folder cypress",
+    "build": "tsc",
+    "build:watch": "tsc --watch"
   },
   "repository": {
     "type": "git",
@@ -37,6 +38,7 @@
     "cypress": "6.9.1",
     "express": "4.17.1",
     "husky": "3.1.0",
+    "typescript": "4.5.4",
     "semantic-release": "17.4.7",
     "start-server-and-test": "1.14.0",
     "stop-only": "3.1.0",
@@ -47,6 +49,7 @@
     "cypress": ">=3"
   },
   "dependencies": {
+    "@types/common-tags": "1.8.1",
     "common-tags": "1.8.0"
   },
   "publishConfig": {
