+++ server-public/test.html
@@ -1,0 +1,1 @@
+MY PAGE
