+++ src/index.ts
@@ -1,0 +1,2 @@
+import * as support from "./support";
+export { support };
